소스코드 정리 - 정경희
