+++ CodeGreen/src/main/java/com/greener/codegreen/dao/BoardDAO.java
@@ -23,6 +23,8 @@ public interface BoardDAO {
 		//FAQ 등록 처리
 		public int FaqUpForm(BoardDTO boardDTO);
 		
+	
+		
 //----------------------------------------------------------------------------	
 // 공지사항, FAQ ,1:1 목록보기 
 //----------------------------------------------------------------------------
