+++ CodeGreen/src/main/java/com/greener/codegreen/service/BoardService.java
@@ -19,6 +19,7 @@ public interface BoardService {
 	//FAQ 작성등록
 	public int FaqUpForm(BoardDTO boardDTO) throws Exception;
 	
+	
 //----------------------------------------------------------------------------	
 // 공지사항, FAQ ,1:1 목록보기 
 //----------------------------------------------------------------------------		
@@ -58,9 +59,7 @@ public interface BoardService {
 //----------------------------------------------------------------------------	
 // 공지사항, FAQ ,1:1  페이징처리 - 뺌 
 //----------------------------------------------------------------------------	
-//----------------------------------------------------------------------------
-// 카테고리에 따른 게시판 보여주기 
-//----------------------------------------------------------------------------
+
