+++ CodeGreen/src/main/resources/mappers/boardMapper.xml
@@ -7,10 +7,6 @@
 
 
 	
-	
-	 <!-- *********************************************************************************** -->
-	 <!-- 공지사항,FAQ 게시물 작성(등록) : 기준이 되는 번호 가져오기  :  번호 1 해주기 위해서 생성되었던것임     -->
- 	 <!-- *********************************************************************************** -->
 
 	 <!-- *********************************************************************************** -->
 	 <!-- 공지사항 게시물 작성(등록) : 이전에 있던 게시물 넣어주기  -->
@@ -23,9 +19,7 @@
 	   		
 	   ]]>
 	</insert>	
- 	 <!-- *********************************************************************************** -->
-	 <!-- 공지사항 게시물 작성(등록) : 새로운 글 등록해주기 이때 글이 등록되는거는 +1 해줌  -->
- 	 <!-- *********************************************************************************** -->
+
 	<!-- *********************************************************************************** -->
 	 <!-- FAQ 게시물 작성(등록) : 이전에 있던 게시물 넣어주기    -->
  	 <!-- *********************************************************************************** -->
@@ -36,15 +30,12 @@
 	   		VALUES (#{f_no}, #{f_bc_code}, #{f_m_id}, #{f_title}, #{f_content} , now() , 0) ;
 	   ]]>
 	</insert>	
- 	 <!-- *********************************************************************************** -->
-	 <!-- FAQ 게시물 작성(등록) : 새로운 글 등록해주기 이때 글이 등록되는거는 +1 해줌  -->
- 	 <!-- *********************************************************************************** -->
-	
+ 
 	
 	
 	
 	 <!-- *********************************************************************************** -->
-	 <!-- 공지사항, FAQ , 1:1문의 (목록 조회 )  + 카테고리 번호에 해당하는 해당 카테고리 목록 조회하기         -->
+	 <!-- 공지사항, FAQ , 1:1문의 (목록 조회 )  + 카테고리 번호에 해당하는 해당 카테고리 목록 조회하기 -->
  	 <!-- *********************************************************************************** -->
 	<select id="Bnoticelist" parameterType="int" resultType="com.greener.codegreen.dto.BoardDTO">
 	<![CDATA[
